20180129新增代码

+++ src/com/alfred/Sailfish/app/Controller/Card/CardRemove.java
@@ -7,8 +7,11 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import com.alfred.Sailfish.app.Service.CardService;
 import com.alfred.Sailfish.app.Util.MethodTool;
+import com.alfred.Sailfish.app.Util.Reference;
 
 /**
  * Servlet implementation class RemoveCard
@@ -30,12 +33,17 @@ public class CardRemove extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		PrintWriter out = initResponse(resp);
-		resp.setCharacterEncoding("utf-8");
-		long card_id = Long.parseLong(req.getParameter("card_id"));
-		long shopmember_id = Long.parseLong(req.getParameter("shopmember_id"));
-		String str = cardService.removeCard(card_id, shopmember_id);
-		System.out.println(MethodTool.getTime() +  ",Response:" + str);
-		out.append(str);
+		HttpSession session = req.getSession(false);
+		if (session == null) {
+			out.append(Reference.SESSION_EXPIRED);
+		} else {
+			resp.setCharacterEncoding("utf-8");
+			long card_id = Long.parseLong(req.getParameter("card_id"));
+			long shopmember_id = MethodTool.getSessionValueToLong(session,"sm_id");
+			String str = cardService.removeCard(card_id, shopmember_id);
+			System.out.println(MethodTool.getTime() +  ",Response:" + str);
+			out.append(str);
+		}
 	}
 	
 	private PrintWriter initResponse(HttpServletResponse resp) {
