+++ src/com/alfred/Sailfish/app/Controller/Classroom/ClassroomRemove.java
@@ -8,8 +8,11 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import com.alfred.Sailfish.app.Service.ClassroomService;
 import com.alfred.Sailfish.app.Util.MethodTool;
+import com.alfred.Sailfish.app.Util.Reference;
 
 /**
  * Servlet implementation class ClassroomDelete
@@ -32,11 +35,16 @@ public class ClassroomRemove extends HttpServlet {
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		resp.setCharacterEncoding("utf-8");
 		PrintWriter out = resp.getWriter();
-		long cr_id = MethodTool.reqParseToLong(req, "cr_id");
-		String str = classroomService.remove(cr_id);
-		System.out.println(MethodTool.getTime() +  ",Response:" + str);
+		HttpSession session = req.getSession(false);
+		if (session == null) {
+			out.append(Reference.SESSION_EXPIRED);
+		}else {
+			long cr_id = MethodTool.reqParseToLong(req, "cr_id");
+			String str = classroomService.remove(cr_id);
+			System.out.println(MethodTool.getTime() +  ",Response:" + str);
+			out.append(str);
+		}
 
-		out.append(str);
 	}
 
 	/**
