+++ src/com/alfred/Sailfish/app/Controller/MemberCard/MemberCardRemove.java
@@ -8,9 +8,11 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.alfred.Sailfish.app.Service.MemberCardService;
 import com.alfred.Sailfish.app.Util.MethodTool;
+import com.alfred.Sailfish.app.Util.Reference;
 
 /**
  * Servlet implementation class MemberCardRemove
@@ -33,11 +35,16 @@ public class MemberCardRemove extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		response.setCharacterEncoding("utf-8");
 		PrintWriter out = response.getWriter();
-		long mc_id = Long.parseLong(request.getParameter("mc_id"));
-		long sm_id = Long.parseLong(request.getParameter("sm_id"));
-		String str = memberCardService.remove(mc_id, sm_id);
-		System.out.println(MethodTool.getTime() +  ",Response:" + str);
-		out.append(str);
+		HttpSession session = request.getSession(false);
+		if (session == null){
+			out.append(Reference.SESSION_EXPIRED);
+		}else {
+			long mc_id = Long.parseLong(request.getParameter("mc_id"));
+			long sm_id = MethodTool.getSessionValueToLong(session,"sm_id");
+			String str = memberCardService.remove(mc_id, sm_id);
+			System.out.println(MethodTool.getTime() +  ",Response:" + str);
+			out.append(str);
+		}
 	}
 
 	/**
