+++ src/com/alfred/Sailfish/app/Controller/Shopmember/ShopmemberDelete.java
@@ -2,13 +2,16 @@ package com.alfred.Sailfish.app.Controller.Shopmember;
 
 import com.alfred.Sailfish.app.Service.ShopmemberService;
 import com.alfred.Sailfish.app.Util.MethodTool;
+import com.alfred.Sailfish.app.Util.Reference;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
+import java.io.PrintWriter;
 
 @WebServlet(name = "ShopmemberDelete",urlPatterns = "/ShopmemberDelete")
 public class ShopmemberDelete extends HttpServlet {
@@ -17,11 +20,18 @@ public class ShopmemberDelete extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setCharacterEncoding("utf-8");
-        long sm_id = MethodTool.reqParseToLong(request,"id");
-        long lmu_id = MethodTool.reqParseToLong(request,"lmu_id");
-        String str = shopmemberService.deleteShopmember(sm_id,lmu_id);
-        System.out.println(MethodTool.getTime() +  ",Response:" + str);
-        response.getWriter().append(str);
+        PrintWriter out = response.getWriter();
+        HttpSession session = request.getSession(false);
+        if (session == null) {
+            out.append(Reference.SESSION_EXPIRED);
+        }else {
+            long sm_id = MethodTool.reqParseToLong(request,"id");
+            long lmu_id = MethodTool.getSessionValueToLong(session,"sm_id");
+            String str = shopmemberService.deleteShopmember(sm_id,lmu_id);
+            System.out.println(MethodTool.getTime() +  ",Response:" + str);
+            response.getWriter().append(str);
+        }
+
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
